requirements.txt
Django==1.7.1
psycopg2==2.5.4

django-debug-toolbar==1.2.2
django-extensions==1.4.6
django-grappelli==2.6.2
python-social-auth==0.2.1
django-recaptcha==1.0.2
django-braces==1.4.0
django-crispy-forms==1.4.0
crispy-forms-foundation==0.4
django-zurb-foundation==5.4.7.1
djangorestframework==3.0.2
Markdown==2.5.2
django-filter==0.9.1
